requirements.txt
async-timeout==4.0.2
bluepy==1.3.0
defusedxml==0.7.1
ifaddr==0.2.0
numpy==1.23.3
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
Pillow==9.2.0
pyparrot==1.5.20
untangle==1.2.1
zeroconf==0.39.1
